Generate provider blocks with variables in root module
User provider blocks in NIA configs are parsed and templated to
the root module. It consumes from variables that are TODO from
variables.tf file. Adds task description as a comment above
the module block for a task to increase context. Adds a preamble
warning that the file may be clobbered.