change e2e to not require sudo
Instead of downloading terraform to /usr/loca/bin, just download it to
the local working directory. Terraform is called using its absolute path
by the client code so it doesn't need to be on the system's searched
PATH.

Also (.git)ignore the downloaded binary and have `make clean` remove it.