change readwrite to use 1-to-1 driver/task setup
This refactoring of the readwrite controller to have 1 driver per
task (expandable to per workflow in future). Having 1 task per driver
means we can use the existing driver API for running tasks separately.
It also lessens the drivers importance as it is now strictly used to
delegate the functionality for the task instead of defining it for all
tasks.

The point of this is to allow executing tasks individually as that is
required for updates (monitor based individual taks runs).