readwrite controller to process tasks in parallel
Refactor code so each unit is handled in a separate goroutine. Includes
checking its template with the resolver and if new/updated it renderes
the template and runs the tf-exec all in that goroutine.

It synchronizes things at the end so the call to process all the tasks
ends when they are all done. This is needed for watcher.Wait() as it
clears the changed data from the last pass so it can track the new
incoming changes.