feature to run through everything once and exit
Add a Once mode to running the readwrite controller. It should run
through everything one time, rendering all templates and executing the
tasks then exit.

It doesn't do anything in parallel at this point to make it better for
debugging. This might require some additional handling at some point as
that means the first pass of running things will take longer than all
the others.

This is very useful for testing as well as adding a flexibility for
users who want to run it in a 'batch' mode.