following once, daemon mode starts off waiting
As we run completely through all tasks an intitial time in the Once
block, we shouldn't immediately re-run through the tasts when kicking
off daemon-mode/Run(). This changes the Run's loop to first Wait on
updates/changes then run the tasks. This will add a Wait break between
the runs, the behavior we want.

Note this doesn't impact the actual working of the system that much as
each run checks to see if that template/task's data was updated and
skips it if not. But it does read weird in the logs and we don't want to
confuse users right off.

We should revisit this down the line once async tasks are here.