Update handlers to return errors
We added support for a ‘once’ mode which, when running the daemon, expects a successful initial cycle of the daemon and will stop if errors are returned. Handlers originally did not return  errors in order to prevent those errors from stopping the daemon. With ‘once’ mode, handlers should be updated to return errors to stop the daemon during ‘once’ mode and just be logged afterwards.

Changes:
 - Update handler.Do interface to return error
 - Update fake & panos handler implementations
 - Pass errors down a chain of handlers as wrapped errors