Fix extra comma in tfvars when no instances of a service
When there are no instances of a service, this can lead to a dangling comma in the services block of `terraform.tfvars`. For example, when there are no instances of any service, the services block will look something like `services = {,}`. Similarly if some services do not have instances, it can lead to an extra preceding/succeeding comma. These extra commas occasionally lead to an invalid expression error when doing a terraform-apply.

Change: add a comma template that determines whether to render a comma depending on the existence of instances for the preceding and succeeding service.