Update handlers to triggered per/task
This is an improvement to trigger handlers at the driver-layer rather than the controller-layer. Previously handlers for all tasks were retrieved at the controller-layer so that all handlers were triggered after each task completed. This leads to handlers unnecessarily being triggered by unrelated tasks. This changes retrieves handlers for a given task and stores it on the task’s driver.