Capture data on process of updating network infrastructure in `Event`
First part in a series of changes to support a status api. This captures data in
an `Event` struct each time CTS goes through the process needed for updating
network infrastructure for a task. Future changes will store the event data and
return various aggregations of the data at the api endpoint

Changes
 - Create `event` package & structure
 - Consume event in controller checkApply()
 - Update `unit` with provider, service, and source in order to configure `event`
 - Update `Task` with ProviderNames() & ServicesNames to help configure `event`