add test mode to retry for testing
Retry in test mode has a fixed delay of a nanosecond. For cases where
you want to verify retries are done but you don't need the actual delay.

Note I did it as a test-mode on the Retry struct instead of switching it
to a Retrier interface as I wanted the nil/zero Retry to work, which it
does, where the interface causes panics and requires nil tests (blah).

I also changed the methods from pointer receivers to values as the
struct values are not changed in place by any methods and so we don't
need the overhead of the pointer reciever.