Attempt to improve consul service templating flakiness
TestRenderTFVarsTmpl/happy_path occasionally fails where the templated service
information is missing one or two of the registered services. Of the 4 fails
observed, two are missing api.worker-01.dc1 and api-2.worker-01.dc1. The other
two are missing web.worker-01.dc1

I haven’t been able to debug this successfully (it started passing when I tried).
This change a pre-emptive measure to increase the context cancel a little and
retry the test once in case the race is between the test and the services
registering.