Refactor map of drivers to a Drivers object
A map of drivers is shared across the API handlers and read-write controller
to manage task lifecycle. This map of drivers is currently used with read-access
only. However, we will want to support write-access in order to support a
create task cli and delete task cli.

Drivers structure will wrap the map of drivers in order to manage concurrent
read/write access to the map.