workaround consul/sdk's TestServer benchmark issue
Several TestServer methods still require *testing.T as the argument
which makes them unusable from a benchmark. This copies the one
remaining method we use (and it's supporting code) from the sdk into
testutils/sdk/. The only changes made were to change the *testing.T to
testing.TB and the smallest transformation possible to convert things
from private methods to local functions.

Once a new version of Consul's SDK is released (current is 0.7.0) we can
dump this and switch back as they've already excepted a PR to fix the
*testing.T issue and we are just awaiting the (unscheduled) release.