move bulk consul service loading to testutils
Move the functions used to load consul with test servers/instances into
the testutils module for reuse in other tests.

Also moved the ShowMe* functions as they are also generally handy.

Cleaned up the loading functions a bit and refactored the existing tests
to use the shared functions instead of the old, local versions.