Update e2e tests to poll API availability
The sleep times kept having to be increased. It is sensitive to the
load of the machine and changes to the underlying CTS logic. The
new CTS client method allows the test to configure the max timeout
waiting for the API to start up and not have to wait the full duration.
The API availability determines that CTS has completed executing the
tasks once through successfully.