fix inspect mode log output
Inspect mode was being logged as daemon mode. In report it was requested
to be reported as inspect mode, so went with that.

This also collects the log output line for each mode into 1 place. To
make it easier to find and more consistency in when mode log line is output.