Rename catalog-services condition 'enable_tf_var' to 'source_includes_var'
The 'enable_tf_var' field is a boolean that should be set based on whether or
not the task's configured source module includes the catalog_services variable
in the terraform `variables.tf` file.

"enable" suggests that this field is optional. However, the field value isn't
truly optional since it's dependent on the source module. "tf_var" language
seems more module-author friendly, but the person writing the config may be a
different person.

Considerations for "source_includes_var":
 - "source" hints that this field should be configured based on the task's
 configured source
 - "var" abbreviates the "variable" terminology in the source module
 - "includes" describes the module to variable relationship
 - can be interpreted as "does the source include var?" which hopefully
 increases intuition in configuring this field