Add custom hcat template function to filter catalog services by regex
In order to support catalog-service condition, we want to be able to filter
catalog services by regex such that only [de]registration of catalog services
with names that match the configured regex trigger the task. The current state
of this feature will trigger the task when a service name does not match the
regex (though the task will make no changes when triggered).

This PR adds a custom hcat template function to do this custom regex filtering
on the catalog service api results. In the next PR, this custom template func
will also handle tag trigger suppression, so changes for this future feature are
anticipated in this current pr.

Changes:
 - Added template function `catalogServicesRegistration`
 - Copy over some helper methods used to create template funcs from hcat
 - Copy and update similar `services` template func tests. Updated for regex.
 Will have e2e tests that will cover Fetch() rather than building out a test
 client system.