Relax task.services configuration requirement depending on condition config
Tasks with a catalog-service condition configured may not need to monitor
service instances, which is configured in task.services. When a catalog-service
condition regexp field is configured, the task monitors and triggers on matches
to the regexp. When catalog-service condition regexp field is not configured,
the regexp value defaults to matching on the task.services values.

Therefore, we want to relax the task.services config requirement only when
catalog-service regexp is configured.

Changes:
 - Update TaskConfig Validate() to check condition config when task.services is
 empty
 - Update CatalogServicesConditionConfig Validate() and Finalize() to handle nil
 regexp and when task.services is empty