Resolve vault dep conflict (#285)
Remove dep on vault internals

Vault dependency was clashing with the pre-go mod version of Vault
used by the Terraform dep. The Terraform dep is used for hcl2shim
but is not meant to be imported and so imports the whole project
which caused conflict with the usage of the `vault/http` and `vault/vault`
packages for testing purposes.

The changes update the vault test usage to just the `vault/api` module
and removes the no longer necessary replace block.