Print out log snippet when CTS fails in e2e tests
Sometimes e2e tests fail in CircleCI and are not reproducible locally.
Currently no e2e logs are outputted in Circle since the tests are run in
parallel making the logs interwoven and undecipherable.

Change: when CTS stops and tests have failed, output a snippet of the CTS logs.
Update e2e default log-level to DEBUG since we will likely be debugging when
these logs are outputted.