Add WaitForEvent method in e2e tests
Currently our e2e tests will call time.Sleep() in order to wait for CTS to
respond to a change in Consul. This WaitForEvent() will wait for CTS to respond
to a change in Consul by checking the task status for a new event rather than
an arbitrary sleep. WaitForEvent() can be set with a liberal wait timeout.

This will be consumed in a later commit alongside a change to wait for Consul
service registration