Store an event when a task becomes enabled and executes
When a task is disabled and then enabled via CLI, the task may execute if there
are changes. Whenever a task executes, an event should be stored.

When enable CLI was implemented, storing an event on task execution was
unintentionally not implemented. This change fixes this and stores an event.

Changes:
 - Identifies when a task is executed: when ?run=now i.e. runOp == driver.RunOptionNow
 - Stores an event
 - Retrieves an error that may result from driver.UpdateTask to store in the event