Handle template rendering when no dependencies have changed
Because of previous commit, when a task is re-enabled, it will re-use the
existing template. Re-rendering an existing template (vs. rendering a new
template) when a task is reenabled introduces a new issue.

If no dependencies have changed between the time the task was disabled and
re-enabled, an already existing and rendered template will infinitely loop
because result.Complete (below) will always be false

```
for i := int64(0); ; i++ {
	if result, err = tf.resolver.Run(tf.template, tf.watcher); err != nil {
		...
		return err
	}
	if result.Complete {
		if _, err = tf.template.Render(result.Contents); err != nil {
		  ...
		  return err
		}
		break
	}
}
```

Changes:
 - Pull in hcat changes https://github.com/hashicorp/hcat/pull/57: new result.NoChange to distinguish when no dependencies
 have changed
 - In renderTemplate, consume result.NoChange to return that rendering is done