Backport of Fix re-enabled task to trigger into release/0.1.x (#323)
* Update e2e tests to poll API availability

The sleep times kept having to be increased. It is sensitive to the
load of the machine and changes to the underlying CTS logic. The
new CTS client method allows the test to configure the max timeout
waiting for the API to start up and not have to wait the full duration.
The API availability determines that CTS has completed executing the
tasks once through successfully.

* Remove dep on vault internals

Vault dependency was clashing with pre-go mod version of Vault
used by the Terraform dep. The Terraform dep is used for hcl2shim
but is not meant to be imported and so imports the whole project

* Run vault integraiton test in CI

* Update vault dep

* Run vault integration tests to weekly

* Remove k8 replace that is no longer required after vault dep update

* cherry-pick manual mishaps

* Copy HCL2ValueFromConfigValue and remove TF dep

* Update task template only when new template content/ID()

Fixes the issue where re-enabled tasks do not execute on trigger and CTS hangs.

This issue was because when a task is re-enabled, it goes through the update
task process which re-initializes the task template. Re-initializing the task
template leads to creating a new task template that has identical content as
the task template from prior initializations. The problem with templates having
identical content is best described: https://github.com/hashicorp/consul-terraform-sync/pull/167

Solution: only create a new task template if there is new template content. Template content unique-ness can be determined by looking at the template.ID()

Change:
 - Only create a new template for the driver for the new template.ID() is
 different from the existing template.ID()

* Handle template rendering when no dependencies have changed

Because of previous commit, when a task is re-enabled, it will re-use the
existing template. Re-rendering an existing template (vs. rendering a new
template) when a task is reenabled introduces a new issue.

If no dependencies have changed between the time the task was disabled and
re-enabled, an already existing and rendered template will infinitely loop
because result.Complete (below) will always be false

```
for i := int64(0); ; i++ {
	if result, err = tf.resolver.Run(tf.template, tf.watcher); err != nil {
		...
		return err
	}
	if result.Complete {
		if _, err = tf.template.Render(result.Contents); err != nil {
		  ...
		  return err
		}
		break
	}
}
```

Changes:
 - Pull in hcat changes https://github.com/hashicorp/hcat/pull/57: new result.NoChange to distinguish when no dependencies
 have changed
 - In renderTemplate, consume result.NoChange to return that rendering is done

* Cleanup old template from hcat watcher when setting new template

* Fix UpdateTask to enable a disabled task that has never run

Fixes the failing tests that were waiting on the wrong
event, and prematurely continuing the test before the catalog
changes were realized as Terraform changes.

* Add WaitForEvent method in e2e tests

Currently our e2e tests will call time.Sleep() in order to wait for CTS to
respond to a change in Consul. This WaitForEvent() will wait for CTS to respond
to a change in Consul by checking the task status for a new event rather than
an arbitrary sleep. WaitForEvent() can be set with a liberal wait timeout.

This will be consumed in a later commit alongside a change to wait for Consul
service registration

* Update RegisterConsulService to wait and poll for service registrations

Most of our e2e tests with call time.Sleep() after calling
RegisterConsulService. Rather than wait an arbitrary amount of time, add polling
in RegisterConsulService to wait and poll GET '/agent/service/:id' until the
service is registered. RegisterConsulService() can be set with a liberal wait
timeout.

This will be consumed in a later commit alongside WaitForEvent()

Changes:
 - Move existing RegisterConsulService from utils.go => consul.go
 - Add new function to check if a service is registered
 - Update RegisterConsulService() to register service and then poll to see if
 service is registered

* Clean-up some e2e tests' time.Sleep()

Changes:
 - Add new variables for default wait times to make it easier to adjust all
 tests when machine is slow: defaultWaitForRegistration, defaultWaitForEvent,
 defaultWaitForNoEvent
 - Update calls to RegisterConsulService() with new polling timeout
 - Add WaitForEvent() after RegisterConsulService() for tests that expect CTS
 to be triggered
 - Refactor time.Sleep() to use defaultWaitForNoEvent for tests that do _not_
 expect CTS to be triggered
 - Update a Deregistration (only 1) to WaitForEvent a little longer:
 defaultWaitForRegistration + defaultWaitForEvent

* Update non-e2e tests with latest RegisterConsulService()

Update these tests with new timeout parameter for RegisterConsulService()

* Add a default timeout variable for cts.WaitForAPI()

Added a new default variable to be used for the amount of time to wait for
WaitForAPI. This makes adjusting the timeout across tests easier when the
machine running the tests is slow. Increased the default value from 15s to 20s
since there seem to be ocasional timeouts.

* Remove extra MakeTempDir call from e2e test

* manual cherry-pick fixes

Co-authored-by: findkim <6362111+findkim@users.noreply.github.com>
Co-authored-by: Lorna Song <lorna@hashicorp.com>