Backport of Fix re-enabled task to trigger into release/0.2.x (#324)
* Update task template only when new template content/ID()

Fixes the issue where re-enabled tasks do not execute on trigger and CTS hangs.

This issue was because when a task is re-enabled, it goes through the update
task process which re-initializes the task template. Re-initializing the task
template leads to creating a new task template that has identical content as
the task template from prior initializations. The problem with templates having
identical content is best described: https://github.com/hashicorp/consul-terraform-sync/pull/167

Solution: only create a new task template if there is new template content. Template content unique-ness can be determined by looking at the template.ID()

Change:
 - Only create a new template for the driver for the new template.ID() is
 different from the existing template.ID()

* Handle template rendering when no dependencies have changed

Because of previous commit, when a task is re-enabled, it will re-use the
existing template. Re-rendering an existing template (vs. rendering a new
template) when a task is reenabled introduces a new issue.

If no dependencies have changed between the time the task was disabled and
re-enabled, an already existing and rendered template will infinitely loop
because result.Complete (below) will always be false

```
for i := int64(0); ; i++ {
	if result, err = tf.resolver.Run(tf.template, tf.watcher); err != nil {
		...
		return err
	}
	if result.Complete {
		if _, err = tf.template.Render(result.Contents); err != nil {
		  ...
		  return err
		}
		break
	}
}
```

Changes:
 - Pull in hcat changes https://github.com/hashicorp/hcat/pull/57: new result.NoChange to distinguish when no dependencies
 have changed
 - In renderTemplate, consume result.NoChange to return that rendering is done

* Cleanup old template from hcat watcher when setting new template

* Fix UpdateTask to enable a disabled task that has never run

Fixes the failing tests that were waiting on the wrong
event, and prematurely continuing the test before the catalog
changes were realized as Terraform changes.

Co-authored-by: Lorna Song <lorna@hashicorp.com>
Co-authored-by: findkim <6362111+findkim@users.noreply.github.com>