Main loop skips active drivers running a task
Adds logic to track when a task is actively run by a driver from
an API/CLI request. This avoids the controller's main loop to not
from racing against the API server.

If it were the other way around, the controller is running a task,
then the driver locks will have the API request to wait until it
is completed before continuing processing the request.

This removes the intermediate "unit" struct and a cleanup refactor
to not require the commands package to know about API and events
package.