Stop controller watcher in rendering test
When e2e tests fail, there will be a large number of lines in the logs will be
outputted to consul similar to:

`[DEBUG] (hcat) error connecting with Consul. Waiting 16.90954134s to retry attempt #5`

These lines are outputted from `TestServicesRenderRace` even when it is not
the cause of the e2e tests failing because the controller's watcher has not been
stopped.