Update BufferPeriodConfig.Finalize()
There is a global-level and task-level BufferPeriod configuration where task-
level BufferPeriod config overrides the global-level config for a given task.

This override is currently resolved at the controller-level instead of the
config-level, which leads to issues. At the controller-level, when deciding to
override the global-level with the task-level, it can be difficult to know if
the task-level values are configured by the user or given values by Finalize()

Solution: resolve the global-level vs. task-level BufferPeriod at config-level.
Task-level config is the source of truth for the buffer period

Changes:
 - Pass in new parent parameter to Finalize() to use to finalize a
 BufferPeriodConfig. This allows the global-level to be resolved when the
 task-level config is finalized
 - Update Finalize to set Min and Max to 0 seconds for disabled buffer periods
 (reduces confusion when debugging. doesn't affect functionality since consumers
 check the Enabled field)