Replace keyExists/key combination with keyExistsGet
Having the two dependencies where the second dependency (the get) depends on the
value of the first dependency (if it exists) has made this too much of a special
case and adds too much complexity. Replacing with a single dependency keeps it
more in line with all our other templates.