Register a new node instead of a health check
Adding a new health check creates two events, the second of which is difficult for us
to time when it exactly occurs. The regex test doesn't need to test checks in particular,
so replacing with a simpler use case of adding a second node to the service. If we want
to test checks, we can do so in a separate test.