Create separate TLS config type for CTS API
Also adds verify_incoming  which will be used for mTLS. The existing TLS config
is very similar to what we need for the CTS API, but we don't currently use
verify and server_name and have the additional verify_incoming.

The attributes also have different meanings in the context of TLS on the CTS API
since its from the perspective of a server and not a client.