Remove unused attributes from run response
The status and the changes_present are currently not needed because
a dry run task will only return a run response if inspection is
successful and will always have changes_present set to true.

We may add these attributes back in later, but leaving them out
until they are relevant.