Create event only when there is a successful run
Previously, we were always creating an event. If the task was not created in the end,
we would then have an oprhaned event that had no driver associated with it. This caused
an error when getting the overall event statuses.

For a task that is created and not run, the status will be unknown until the first run
is triggered.