Add auto-approve option to delete and enable
Adds an option to skip the interactive approval step when enabling a task that has
changes or deleting a task. The flag name was based on a similar option for Terraform
apply. In the case of enabling a task, a Terraform plan will still run before the apply
so that users are still able to see what changes will happen.