Add VariableType() to different types of MonitorConfig
Context: this is work to needed to support a new multiple module_input feature,
which requires that tasks cannot have multiple configurations for the same
variable type. That is, module_input blocks must be unique across other
module_inputs, condition, and services.

Adding VariableType() to MonitorConfig and ScheduleCondition will allow us to
check the variable type and confirm uniqueness.