Update notifiers to allow different types of module inputs
Many notifiers New*() constructors have a "serviceCount" parameter because
services was the main module input option.

Now that we are allowing different types of module inputs, this integer
should be a count of the number of monitored tmplfuncs in the template, which
can be more than the number of services.

Updates:
 - Update notifier constructors from "serviceCount" => "tmplFuncCount". This
 eliminates the need of random +1 in the constructor
 - Rename notifiers constructors from "depCount" => "tmplFuncCount" to align
 on naming
 - Update some comments and local variables to be the same across notifiers