Initialize disabled tasks
When a task is initially disabled and the initialization is skipped, the template
is not generated, so there is no mapping of template ID to this task. This caused
an issue when the task is enabled because the controller could not determine what
task to run when it got a change from the template.

Remvoing the restriction of initializing only enabled tasks also allows validation
issues tasks to be caught at creation rather than at enablement.