Rename Template method SourceIncludesVariable() => RendersVar()
The variable that SourceIncludesVariable() returns was renamed from
"SourceIncludesVar" to "RenderVar".

Rename the method to RendersVar() which informs the method consumer whether the
template renders the variable or not.

Note: "RenderVar" (no 's') sounds more of an imperative statement which is
useful for the consumer configuring the Template.

RenderVar vs. RendersVar also avoids a DuplicateFieldAndMethod compile error