Update NoMonitorConfig to NoConditionConfig
Since module_input is now an array of module inputs, there is no need for it to
have this special non-nil empty module input i.e. it can now be an empty array.

NoMonitorConfig is only needed now for non-nil empty condition. Update the type
name to reflect the change from Monitor => Condition