Add delay in rendering template for task creation
The RenderTemplate call is called continuously until the template is complete and logs
a TRACE line every time. This can result in thousands of calls/log lines. Adding a small
delay to reduce this.

This only is needed in API creations because the templates rendered
at startup block for a dependency update before rendering again.