Add basic compile test for weekly tests to CI
Some of our weekly tests have separate build tags.

This makes it easy to accidentally make a change so that they cannot compile
without noticing. Because these tests also run weekly, the issue may not be
discovered until later.

This change adds a compile check to be run in circleci when a branch is pushed
or on a merge to main.