Mark Task Status API response's config-related information as deprecate
In the Task Status API, there is some non-status related information that is
more task configuration details.

We want to deprecate this config-related information. Users will be able to
switch to using Get Task API to retrieve this information

Add comments noting that responses are deprecated. Users might use use these
objects to unmarshall JSON. Leaving as is to prevent a breaking change.