Fix early exit when err is nil for rw/ro Once & Run
This catches an issue with multiple goroutines, if one routine completes
successfully, it sends a nil error to the exit channel. This was incorrectly
causing rw.Once and ro.Run to exit early instead of waiting for the other
goroutine to finish

rw.Run goroutines never complete and therefore never send a nil error. These
goroutines get canceled and therefore send a context.Canceled error. Added a nil
check here just to be safe.