Update WaitForAPI() to check for enabled tasks with unknown tasks
Context: e2e tests rely on WaitForAPI() function to determine when the tasks
have all run once. This used API availability to identify once-completion.

Changes:
 - Rename WaitForAPI => WaitForOnce
 - Change function to identify once-completion by looking at the each task's
 Task Status response to check for enabled tasks that haven't run yet

Note: we are planning to move when the API is served, so no longer can rely on
API availability to determine when once-mode is complete