Register CTS only after once-mode is complete
The test TestE2ERestartConsul uncovered an issue where registering the service
before once-mode has finished causes CTS to ignore a service change after Consul
is restarted. This doesn't fix the root cause, but it will keep it from happening
because of the self-registration.

We were deregistering the service on failure anyway, so it also makes sense to
just register afterwards when we know all tasks have passed.