Remove check from catalog registration
This request (which is not made by CTS) now fails with versions of Consul
earlier than 1.10.0 due to the introduction of the ExposedPort field on
the check. The test itself was only verifying a change in the node
name and not the check name, so removed the check so it would work with
all versions of Consul.