Fix issue with task config incorrectly modified.
Global workdir and buffer periods are no longer tightly coupled with
task configuration. This change is necessary due to the fact that task
configs are marshaled in and out of the API, which means that modifying
these fields is not safe.

Modified task configs are now cloned and tweaked within the driver factory
to ensure that they do not accidentally leak into the API and stores.